% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp2019.R
\docType{data}
\name{atp2019}
\alias{atp2019}
\title{Association of Tennis Professionals 2019 (\code{atp2019})}
\format{
A data frame with 14701 rows and 16 variables:
\describe{
  \item{match_id}{String, game's identifier. Not null.}
  \item{double}{Character, if the column double has the letter t, the game is a double match. Not null.}
  \item{round_number}{Number, 0 represents the final game, 1 the semi-final, and so on. Not null.}
  \item{w1_id}{String, identifier of player 1. Not null.}
  \item{w1_name}{String, name of player 1. Not null.}
  \item{w2_id}{String, identifier of player 2}
  \item{w2_name}{String, name of player 2}
  \item{l1_id}{String, identifier of player 1. Not null.}
  \item{l1_name}{String, name of player 1. Not null.}
  \item{l2_id}{String, identifier of player 2}
  \item{l2_name}{String, name of player 2}
  \item{time_start}{String, format = Y-m-d. Not null.}
  \item{time_end}{String, format = Y-m-d}
  \item{ground}{String, ground type.}
  \item{tour_id}{String, tournament's identifier}
  \item{tour_name}{String, tournament's name}
}
}
\source{
\url{https://github.com/glandfried/tennis_atp/releases/download/atp/history.csv.zip}
}
\usage{
atp2019
}
\arguments{
\item{atp2019}{A database with all the 2019 ATP matches}
}
\description{
A dataset of the ATP 2019, containing 14701 games played by 1306 unique players.
}
\keyword{datasets}
