% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueskillthroughtime.R
\name{lc_print}
\alias{lc_print}
\title{Print list of Gaussian using the python and julia syntax}
\usage{
lc_print(lc.a)
}
\arguments{
\item{lc.a}{List of Gaussians}
}
\value{
No return value, print lists of Gaussian using the python and julia syntax
}
\description{
Print list of Gaussian using the python and julia syntax
}
